requirements.txt
Flask==3.1.0
google-genai
python-dotenv
